Simplify HPCsim!
Store the PRNG associated with the events in the TLS. That way, the user doesn't need any longer to know about the rand context, they just need to call RandU01() to have a PRN which is statistically sound.

The Pi simulation has been simplified taking into account this welcomed changed.