Add a mechanism that allows performing the reduce operation directly during the simulation. This helps avoiding extra binaries writing in case of simple simulations.
A second example simulation has been added to demonstrate this new capability: PiReduce. It performs a Pi compuation (as Pi simulation) and share the hot loop with Pi simulation. But, instead of writing all the results to the disk, it displays the Pi value at the end of the execution.